Ref: moving getMeta() to metadata.ts.
