+++ src/metadata.ts
@@ -1,7 +1,7 @@
 import {z} from './index';
 
 export const metaProp = 'expressZodApiMeta';
-export type MetaProp = typeof metaProp;
+type MetaProp = typeof metaProp;
 
 export type MetaDef<T extends z.ZodTypeAny> = {
   [K in MetaProp]: {
@@ -9,8 +9,8 @@ export type MetaDef<T extends z.ZodTypeAny> = {
     description?: string;
   };
 };
-export type MetaKey = keyof MetaDef<any>[MetaProp];
-export type MetaValue<T extends z.ZodTypeAny, K extends MetaKey> = Readonly<MetaDef<T>[MetaProp][K]>;
+type MetaKey = keyof MetaDef<any>[MetaProp];
+type MetaValue<T extends z.ZodTypeAny, K extends MetaKey> = Readonly<MetaDef<T>[MetaProp][K]>;
 
 type ExampleSetter<T extends z.ZodTypeAny> = (example: z.input<T>) => WithMeta<T>;
 type DescriptionSetter<T extends z.ZodTypeAny> = (description: string) => WithMeta<T>;
@@ -46,6 +46,14 @@ export const hasMeta = <T extends z.ZodTypeAny>(schema: T): schema is WithMeta<T
   }
   return typeof schema._def[metaProp] === 'object' && schema._def[metaProp] !== null;
 };
+
+export function getMeta<T extends z.ZodTypeAny, K extends MetaKey>(schema: T, meta: K): MetaValue<T, K> | undefined {
+  if (!hasMeta(schema)) {
+    return undefined;
+  }
+  const def = schema._def as MetaDef<T>;
+  return meta in def[metaProp] ? def[metaProp][meta] : undefined;
+}
 
 export const copyMeta = <A extends z.ZodTypeAny, B extends z.ZodTypeAny>(src: A, dest: B): B | WithMeta<B> => {
   if (!hasMeta(src)) {
